requirements.txt
fastapi
uvicorn[standard]

pypdf
pymupdf

streamlit
requests

pinecone
chromadb
langchain-pinecone
langchain-chromadb

langchain
langchain-community
langchain-huggingface
sentence-transformers
huggingface-hub

python-multipart
python-dotenv